% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreAssessmentResults.R
\name{ScoreAssessmentResults}
\alias{ScoreAssessmentResults}
\title{Score Gene Assessment Results}
\usage{
ScoreAssessmentResults(x, mode = "a")
}
\arguments{
\item{x}{An object of class \code{Assessment} and subclass \code{Results}}

\item{mode}{Must either be "a" (use all evidence), "p" (use proteomics evidence only),
or "c" (use evolutionary conservation evidence only)}
}
\description{
Scores the results from the assessment of a set of genes using one of three modes
}
\examples{

ScoreAssessmentResults(readRDS(system.file("extdata", "MGAS5005_PreSaved_ResultsObj_Prodigal.rds", package = "AssessORF")), "a")

ScoreAssessmentResults(readRDS(system.file("extdata", "MGAS5005_PreSaved_ResultsObj_Prodigal.rds", package = "AssessORF")), "c")

ScoreAssessmentResults(readRDS(system.file("extdata", "MGAS5005_PreSaved_ResultsObj_Prodigal.rds", package = "AssessORF")), "p")

}
\seealso{
\code{\link{Assessment-class}}
}
