% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assessment-class.R
\name{plot.Assessment}
\alias{plot.Assessment}
\title{Plot Assessment Objects}
\usage{
\method{plot}{Assessment}(x, y = NULL, related_MinConStart = 0.8, ...)
}
\arguments{
\item{x}{An object of class \code{Assessment} and of either subclass \code{DataMap} or subclass \code{Results}.}

\item{y}{An optional object of class \code{Assessment} and of either subclass \code{DataMap} or subclass \code{Results}. Its
subclass must be different than the subclass of \code{x}}

\item{related_MinConStart}{Minimum value of the conservation to coverage ratio needed to call a start conserved. Must range
from 0 to 1. Lower values allow more conserved starts through. Recommended to use default value.}

\item{...}{Further plotting parameters.}
}
\value{
Invisibly returns the input object \code{x}
}
\description{
The \code{plot} method for \code{Assessment} objects
}
\details{
If only \code{x} is specified and \code{x} is of subclass \code{DataMap}, an interactive genome viewer showing how the
proteomics data and evolutionary conservation data maps to the central genome is plotted.

If only \code{x} is specified and \code{x} is of subclass \code{Results}, a bar chart describing the number of genes in each
category is plotted. For the predicted gene categories, bars are colored by the correctness of that category, where dark green
represents "definitely correct", light green represents "likely correct", white represents "no evidence", dark red represents
"definitely incorrect", light red represents "likely incorrect", and grey represents "potentially incorrect". For the two
categories that come from ORFs without predicted genes, dark blue represents "likely missing" and light blue represents
"potentially missing".

If both \code{x} and \code{y} are specified, an interactive genome viewer showing how the proteomics data, evolutionary
conservation data, and gene set map to the central genome is plotted.
}
\examples{

currMapObj <- readRDS(system.file("extdata",
                                  "MGAS5005_PreSaved_DataMapObj.rds",
                                  package = "AssessORF"))

currResObj <- readRDS(system.file("extdata",
                                  "MGAS5005_PreSaved_ResultsObj_Prodigal.rds",
                                  package = "AssessORF"))

plot(currMapObj)

plot(currResObj)

plot(currMapObj, currResObj)

plot(currResObj, currMapObj)

}
\seealso{
\code{\link{Assessment-class}}, \code{\link{locator}}
}
