% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreAssessmentResults.R
\name{ScoreAssesmentResults}
\alias{ScoreAssesmentResults}
\title{Score Gene Assessment Results}
\usage{
ScoreAssesmentResults(x, mode = "a")
}
\arguments{
\item{x}{An object of class \code{Assessment} and subclass \code{Results}}

\item{mode}{Must either be "a" (use all evidence), "p" (use proteomics evidence only),
or "c" (use evolutionary conservation evidence only)}
}
\description{
Scores the results from the assessment of a set of genes using one of three modes
}
\examples{

ScoreAssessmentResults(AssessORFData::GetResultsObject("ATCC17978"), "a")
ScoreAssessmentResults(AssessORFData::GetResultsObject("ATCC17978"), "c")
ScoreAssessmentResults(AssessORFData::GetResultsObject("ATCC17978"), "p")

}
\seealso{
\code{\link{Assessment-class}}
}
